Changed simulation output into its own module.
